backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
nltk
scikit-learn
bcrypt
requests